diagnosis/pyproject.toml
[project]
name = "diagnosis"
version = "0.1.0"
description = "diagnosis using crewAI"
authors = [{ name = "Jaimin chandarana", email = "jdchandarana3@gmail.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]>=0.159.0,<1.0.0",
    "pytesseract",
    "pillow",
    "PyPDF2"
]

[project.scripts]
diagnosis = "diagnosis.main:run"
run_crew = "diagnosis.main:run"
train = "diagnosis.main:train"
replay = "diagnosis.main:replay"
test = "diagnosis.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
